Soluciona duplicidades detectadas por Sonarcloud
