+++ src/routes/api/newsletters/api.js
@@ -1,0 +1,28 @@
+import { getHeaders } from "$lib/utils";
+
+export const api = async ({request, locals, method, error}) => {
+    const headers = getHeaders(request);
+    const body = await request.json();
+    const apiURI = locals.apiURI;
+
+    const response = await fetch(`${apiURI}/api/admin/newsletters`, {
+        method,
+        headers,
+        body: JSON.stringify(body)
+    });
+
+    if (response.ok) {
+        return {
+            status: 201,
+            headers: response.headers,
+            body: await response.json()
+        }
+    } else {
+        return {
+            status: 500,
+            body: JSON.stringify({
+                message: error
+            })
+        }
+    }
+}
