+++ src/routes/api/newsletters/publicar.js
@@ -1,28 +1,10 @@
-import { getHeaders } from "$lib/utils";
+import { api } from "./api";
 
 export const post = async ({ request, locals }) => {
-    const headers = getHeaders(request);
-    const body = await request.json();
-    const apiURI = locals.apiURI;
-
-    const response = await fetch(`${apiURI}/api/admin/newsletters`, {
+    return await api({
+        request,
+        locals,
         method: "POST",
-        headers,
-        body: JSON.stringify(body)
+        error: "No se ha podido publicar la newsletter con estos datos."
     });
-
-    if (response.ok) {
-        return {
-            status: 201,
-            headers: response.headers,
-            body: await response.json()
-        }
-    } else {
-        return {
-            status: 500,
-            body: JSON.stringify({
-                message: "No se ha podido publicar la newsletter con estos datos."
-            })
-        }
-    }
 }
